+++ composer.json
@@ -31,9 +31,6 @@
         "yiisoft/aliases": "^3.0@dev",
         "yiisoft/di": "^3.0@dev"
     },
-    "provide": {
-        "yiisoft/yii-bootstrap-implementation": "3.0.0"
-    },
     "autoload": {
         "psr-4": {
             "Yiisoft\\Yii\\Bootstrap4\\": "src"
